change uint64 to uint (#130)

+++ client/agent.go
@@ -21,7 +21,7 @@ func (c *Client) RegisterAgent(ctx context.Context, payload types.RegisterAgent)
 func (c *Client) Agents(ctx context.Context, projectID string, params types.AgentsParams) (types.Agents, error) {
 	q := url.Values{}
 	if params.Last != nil {
-		q.Set("last", strconv.FormatUint(*params.Last, uintBase))
+		q.Set("last", strconv.FormatUint(uint64(*params.Last), uintBase))
 	}
 	if params.Before != nil {
 		q.Set("before", *params.Before)
