+++ client/membership.go
@@ -13,7 +13,7 @@ import (
 func (c *Client) Members(ctx context.Context, projectID string, params types.MembersParams) (types.Memberships, error) {
 	q := url.Values{}
 	if params.Last != nil {
-		q.Set("last", strconv.FormatUint(*params.Last, uintBase))
+		q.Set("last", strconv.FormatUint(uint64(*params.Last), uintBase))
 	}
 	if params.Before != nil {
 		q.Set("before", *params.Before)
